Init factorisation
