+++ test/minister_date_integrity.py
@@ -8,50 +8,17 @@ for the same person and role, and intersect the date interval of the referenced
 government in `data/government.csv`. 
 """
 from pathlib import Path
-import re
 import unittest
 
 import pandas as pd
 from trainerlog import get_logger
 
+from test.date_integrity_helpers import parse_date_interval
+
 
 LOGGER = get_logger("minister-date-integrity")
 DATA_DIR = Path("data")
 RESULT_DIR = Path("test/result")
-DATE_RE = re.compile(r"^(\d{4})(?:-(\d{2})(?:-(\d{2}))?)?$")
-
-
-def parse_date_interval(value, is_end):
-    if pd.isna(value) or str(value).strip() == "":
-        if is_end:
-            return pd.Timestamp.max.normalize(), None, "blank"
-        return None, None, "blank"
-
-    value = str(value).strip()
-    match = DATE_RE.match(value)
-    if match is None:
-        return None, None, "malformed"
-
-    year = int(match.group(1))
-    month = int(match.group(2)) if match.group(2) else None
-    day = int(match.group(3)) if match.group(3) else None
-
-    try:
-        if month is None:
-            if is_end:
-                return pd.Timestamp(year + 1, 1, 1), "year", None
-            return pd.Timestamp(year, 1, 1), "year", None
-
-        if day is None:
-            if is_end:
-                if month == 12:
-                    return pd.Timestamp(year + 1, 1, 1), "month", None
-                return pd.Timestamp(year, month + 1, 1), "month", None
-            return pd.Timestamp(year, month, 1), "month", None
-
-        return pd.Timestamp(year, month, day), "day", None
-    except ValueError:
-        return None, None, "malformed"
 
 
 def add_date_columns(df):
